[ feature: issue-26 ] Import Avatars to one file and restructure them into an array of objects
