+++ packages/client/src/assets/mocks/avatarImages.ts
@@ -1,11 +1,13 @@
-import Avatar1 from './avatar1.jpg';
-import Avatar2 from './avatar2.jpg';
-import Avatar3 from './avatar3.jpg';
-import Avatar4 from './avatar4.jpg';
-import Avatar5 from './avatar5.jpg';
-import Avatar6 from './avatar6.jpg';
-import Avatar7 from './avatar7.jpg';
-import Avatar8 from './avatar8.jpg';
-import Avatar9 from './avatar9.jpg';
+import { Avatar1, Avatar2, Avatar3, Avatar4, Avatar5, Avatar6, Avatar7, Avatar8, Avatar9 } from '.';
 
-export const avatarImages = [Avatar1, Avatar2, Avatar3, Avatar4, Avatar5, Avatar6, Avatar7, Avatar8, Avatar9];
+export const avatarImages = [
+  { id: 1, imgPath: Avatar1 },
+  { id: 2, imgPath: Avatar2 },
+  { id: 3, imgPath: Avatar3 },
+  { id: 4, imgPath: Avatar4 },
+  { id: 5, imgPath: Avatar5 },
+  { id: 6, imgPath: Avatar6 },
+  { id: 7, imgPath: Avatar7 },
+  { id: 8, imgPath: Avatar8 },
+  { id: 9, imgPath: Avatar9 },
+];
